// src/modules/cms/transaction/dto/create_transaction_detail.dto.ts
export class CreateTransactionDto {
    vnp_Amount: number | string;

    vnp_BankCode: string;

    vnp_BankTranNo: string;

    vnp_CardType: string;

    vnp_OrderInfo: string;

    vnp_ResponseCode: string;

    vnp_TransactionNo: string;

    vnp_TransactionStatus: string;

    vnp_TxnRef: string;

    vnp_SecureHash: string;

    vnp_PayDate: string;
}